x64 part 13: port Windows sockets callbacks device APIs and dependency selection

+++ src/win32/win_net.h
@@ -22,7 +22,7 @@ void		NET_SendPacket(netsrc_t sock, int length, const void* data, netadr_t to);
 const char* NET_ErrorString(void);
 void		NET_Sleep(int msec);
 
-uint __cdecl NET_TCPIPSocket(const char *net_interface, int port, int type);
+uintptr_t __cdecl NET_TCPIPSocket(const char *net_interface, int port, int type);
 
 qboolean Sys_StringToAdr(const char *s, netadr_t *a);
 char __cdecl Sys_SendPacket(int length, unsigned __int8 *data, netadr_t to);
